mv use inventory
